be_fastapi/pyproject.toml
[project]
name = "be-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "antlr4-python3-runtime==4.9.2",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.121.1",
    "onnxruntime>=1.23.2",
    "onnxruntime-gpu>=1.23.2",
    "onnxscript>=0.5.6",
    "onnxslim>=0.1.74",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.229",
    "uvicorn[standard]>=0.38.0",
    "websockets>=15.0.1",
]
